feat: Shrink (#56)
Add a new "ResizableSize" configuration, labelled "shrink", that will automatically shrink the container to its child size.

Rework how the initial layout is done, allowing Flutter to set the sizes of any "expand" and "shrink" containers. After the initial layout, a routine is executed that measures the sizes of each container and then updates the controller. Any subsequent layouts will use these values and the deltas.

As a personal note, I don't _love_ this implementation, as it requires providing a `GlobalKey` to every child widget and then running an asynchronous process to read the sizes and manually update the controller's values, but I couldn't find a better solution.